Clean up energy/wavelength conversion to single function mkak 2017.01.09
