+++ plugins/diFFit/__init__.py
@@ -4,8 +4,10 @@ from .cifdb import cifDB
 from .XRD2Dviewer import diFFit2D
 from .XRD1Dviewer import diFFit1D,Calc1DPopup
 from .ImageControlsFrame import ImageToolboxFrame
+from .XRDCalculations import d_from_q,d_from_twth,twth_from_d,twth_from_q
+from .XRDCalculations import E_from_lambda,lambda_from_E,q_from_d,q_from_twth
 from .XRDCalculations import integrate_xrd,xy_file_reader,peakfinder,peaklocater
-from .XRDCalculations import calc_q_to_d,calc_q_to_2th,generate_hkl,instrumental_fit_uvw
 from .XRDCalculations import data_gaussian_fit,peakfitter
+from .XRDCalculations import generate_hkl,instrumental_fit_uvw
 from .XRDCalibrationFrame import CalibrationPopup
 from .XRDMaskFrame import MaskToolsPopup
